+++ R/data_interface.R
@@ -1475,8 +1475,8 @@ spark_read <- function(sc,
 #' @export
 spark_write_rds <- function(x, dest_uri) {
   sc <- spark_connection(x)
-  if (spark_version(sc) < "2.0.0") {
-    stop("`spark_write_rds()` is only supported in Spark 2.0 or above")
+  if (spark_version(sc) < "3.0.0") {
+    stop("`spark_write_rds()` is only supported in Spark 3.0 or above")
   }
 
   num_partitions <- sdf_num_partitions(x)